Updated GitHub Workflows

+++ include/arithmetics.h
@@ -11,7 +11,7 @@ double CBP_Arith_RawRepresentationToDouble(int64_t raw_value);
 
 // Note that the two functions are all based on an assumption that HX_CBP_PRECISION == 2.
 // This may be changed later.
-long long CBP_Arith_GetFixedPointRepr(const char*);
+int64_t CBP_Arith_GetFixedPointRepr(const char*);
 
-long long CBP_Arith_GetQuotedFixedPointRepr(int64_t, int64_t);
+int64_t CBP_Arith_GetQuotedFixedPointRepr(int64_t, int64_t);
 #endif
